[ci] move to ValidatorIntegrationTest
