+++ apps/app/src/test/scala/org/lfdecentralizedtrust/splice/integration/tests/ValidatorIntegrationTest.scala
@@ -20,9 +20,16 @@ import com.digitalasset.canton.topology.PartyId
 import com.digitalasset.canton.topology.transaction.ParticipantPermission
 import org.apache.pekko.http.scaladsl.Http
 import org.apache.pekko.http.scaladsl.client.RequestBuilding.{Get, Post}
-import org.apache.pekko.http.scaladsl.model.StatusCodes
+import org.apache.pekko.http.scaladsl.model.{HttpRequest, HttpResponse, StatusCodes}
 import org.apache.pekko.http.scaladsl.model.headers.{Authorization, OAuth2BearerToken}
 import org.lfdecentralizedtrust.splice.config.ConfigTransforms
+import org.lfdecentralizedtrust.tokenstandard.transferinstruction.v1.definitions.GetChoiceContextRequest
+import org.lfdecentralizedtrust.tokenstandard.{
+  allocation,
+  allocationinstruction,
+  metadata,
+  transferinstruction,
+}
 import org.slf4j.event.Level
 
 import scala.concurrent.Future
@@ -720,6 +727,248 @@ class ValidatorIntegrationTest extends IntegrationTestWithIsolatedEnvironment wi
           ).create().commands().asScala.toSeq,
         )
     }
+  }
+
+  "serve all token standard endpoints via the scan-proxy" in { implicit env =>
+    startAllSync(
+      sv1ScanBackend,
+      sv1Backend,
+      sv2ScanBackend,
+      sv2Backend,
+      sv3ScanBackend,
+      sv3Backend,
+      sv4ScanBackend,
+      sv4Backend,
+    )
+
+    aliceValidatorBackend.startSync()
+    onboardWalletUser(aliceValidatorWalletClient, aliceValidatorBackend)
+    val headers = List(
+      Authorization(
+        OAuth2BearerToken(aliceValidatorWalletClient.token.valueOrFail("No token found"))
+      )
+    )
+
+    val scanProxyUrl =
+      s"http://${aliceValidatorBackend.config.adminApi.address}:${aliceValidatorBackend.config.adminApi.port.unwrap}/api/validator/v0/scan-proxy"
+    val send: HttpRequest => Future[HttpResponse] = Http().singleRequest(_)
+    val fakeCid = "00" + "01" * 31 + "42"
+    // Invalid choice arguments: factories are expected to reject them with a 400.
+    val emptyChoiceArgs = io.circe.Json.obj()
+
+    val metadataClient = metadata.v1.Client.httpClient(send, scanProxyUrl)
+    val transferV1Client = transferinstruction.v1.Client.httpClient(send, scanProxyUrl)
+    val transferV2Client = transferinstruction.v2.Client.httpClient(send, scanProxyUrl)
+    val allocationInstructionV1Client =
+      allocationinstruction.v1.Client.httpClient(send, scanProxyUrl)
+    val allocationInstructionV2Client =
+      allocationinstruction.v2.Client.httpClient(send, scanProxyUrl)
+    val allocationV1Client = allocation.v1.Client.httpClient(send, scanProxyUrl)
+    val allocationV2Client = allocation.v2.Client.httpClient(send, scanProxyUrl)
+
+    def check[R](endpoint: String)(response: => Either[Either[Throwable, HttpResponse], R])(
+        expected: PartialFunction[R, Unit]
+    ) =
+      withClue(s"$endpoint via scan-proxy") {
+        inside(response) { case Right(r) => inside(r)(expected) }
+      }
+
+    // We just need to check that the responses can be parsed to know that they're wired correctly
+
+    // metadata v1
+    check("GET /registry/metadata/v1/info")(
+      metadataClient.getRegistryInfo(headers).value.futureValue
+    ) { case metadata.v1.GetRegistryInfoResponse.OK(info) =>
+      info.adminId shouldBe dsoParty.toProtoPrimitive
+    }
+    check("GET /registry/metadata/v1/instruments")(
+      metadataClient.listInstruments(None, None, headers).value.futureValue
+    ) { case metadata.v1.ListInstrumentsResponse.OK(instruments) =>
+      instruments.instruments.map(_.id) shouldBe Vector("Amulet")
+    }
+    check("GET /registry/metadata/v1/instruments/{instrumentId}")(
+      metadataClient.getInstrument("Amulet", headers).value.futureValue
+    ) { case metadata.v1.GetInstrumentResponse.OK(instrument) =>
+      instrument.id shouldBe "Amulet"
+    }
+
+    // transfer-instruction v1
+    check("POST /registry/transfer-instruction/v1/transfer-factory")(
+      transferV1Client
+        .getTransferFactory(
+          transferinstruction.v1.definitions.GetFactoryRequest(emptyChoiceArgs, None),
+          headers,
+        )
+        .value
+        .futureValue
+    ) { case transferinstruction.v1.GetTransferFactoryResponse.BadRequest(_) => }
+    check("POST /registry/transfer-instruction/v1/{id}/choice-contexts/accept")(
+      transferV1Client
+        .getTransferInstructionAcceptContext(fakeCid, GetChoiceContextRequest(None), headers)
+        .value
+        .futureValue
+    ) { case transferinstruction.v1.GetTransferInstructionAcceptContextResponse.NotFound(_) => }
+    check("POST /registry/transfer-instruction/v1/{id}/choice-contexts/reject")(
+      transferV1Client
+        .getTransferInstructionRejectContext(fakeCid, GetChoiceContextRequest(None), headers)
+        .value
+        .futureValue
+    ) { case transferinstruction.v1.GetTransferInstructionRejectContextResponse.NotFound(_) => }
+    check("POST /registry/transfer-instruction/v1/{id}/choice-contexts/withdraw")(
+      transferV1Client
+        .getTransferInstructionWithdrawContext(fakeCid, GetChoiceContextRequest(None), headers)
+        .value
+        .futureValue
+    ) { case transferinstruction.v1.GetTransferInstructionWithdrawContextResponse.NotFound(_) => }
+
+    // transfer-instruction v2
+    check("POST /registry/transfer-instruction/v2/transfer-factory")(
+      transferV2Client
+        .getTransferFactory(
+          transferinstruction.v2.definitions.GetFactoryRequest(emptyChoiceArgs, None),
+          headers,
+        )
+        .value
+        .futureValue
+    ) { case transferinstruction.v2.GetTransferFactoryResponse.BadRequest(_) => }
+    check("POST /registry/transfer-instruction/v2/{id}/choice-contexts/accept")(
+      transferV2Client
+        .getTransferInstructionAcceptContext(
+          fakeCid,
+          transferinstruction.v2.definitions.GetChoiceContextRequest(None),
+          headers,
+        )
+        .value
+        .futureValue
+    ) { case transferinstruction.v2.GetTransferInstructionAcceptContextResponse.NotFound(_) => }
+    check("POST /registry/transfer-instruction/v2/{id}/choice-contexts/reject")(
+      transferV2Client
+        .getTransferInstructionRejectContext(
+          fakeCid,
+          transferinstruction.v2.definitions.GetChoiceContextRequest(None),
+          headers,
+        )
+        .value
+        .futureValue
+    ) { case transferinstruction.v2.GetTransferInstructionRejectContextResponse.NotFound(_) => }
+    check("POST /registry/transfer-instruction/v2/{id}/choice-contexts/withdraw")(
+      transferV2Client
+        .getTransferInstructionWithdrawContext(
+          fakeCid,
+          transferinstruction.v2.definitions.GetChoiceContextRequest(None),
+          headers,
+        )
+        .value
+        .futureValue
+    ) { case transferinstruction.v2.GetTransferInstructionWithdrawContextResponse.NotFound(_) => }
+
+    // allocation-instruction v1
+    check("POST /registry/allocation-instruction/v1/allocation-factory")(
+      allocationInstructionV1Client
+        .getAllocationFactory(
+          allocationinstruction.v1.definitions.GetFactoryRequest(emptyChoiceArgs, None),
+          headers,
+        )
+        .value
+        .futureValue
+    ) { case allocationinstruction.v1.GetAllocationFactoryResponse.OK(_) => }
+
+    // allocation-instruction v2
+    check("POST /registry/allocation-instruction/v2/allocation-factory")(
+      allocationInstructionV2Client
+        .getAllocationFactory(
+          allocationinstruction.v2.definitions.GetFactoryRequest(emptyChoiceArgs, None),
+          headers,
+        )
+        .value
+        .futureValue
+    ) { case allocationinstruction.v2.GetAllocationFactoryResponse.BadRequest(_) => }
+    check("POST /registry/allocation-instruction/v2/{id}/choice-contexts/accept")(
+      allocationInstructionV2Client
+        .getAllocationInstructionAcceptContext(
+          fakeCid,
+          allocationinstruction.v2.definitions.GetChoiceContextRequest(None),
+          headers,
+        )
+        .value
+        .futureValue
+    ) { case allocationinstruction.v2.GetAllocationInstructionAcceptContextResponse.BadRequest(_) =>
+    }
+    check("POST /registry/allocation-instruction/v2/{id}/choice-contexts/withdraw")(
+      allocationInstructionV2Client
+        .getAllocationInstructionWithdrawContext(
+          fakeCid,
+          allocationinstruction.v2.definitions.GetChoiceContextRequest(None),
+          headers,
+        )
+        .value
+        .futureValue
+    ) {
+      case allocationinstruction.v2.GetAllocationInstructionWithdrawContextResponse.BadRequest(_) =>
+    }
+
+    // allocation v1
+    check("POST /registry/allocations/v1/{id}/choice-contexts/execute-transfer")(
+      allocationV1Client
+        .getAllocationTransferContext(
+          fakeCid,
+          allocation.v1.definitions.GetChoiceContextRequest(None),
+          headers,
+        )
+        .value
+        .futureValue
+    ) { case allocation.v1.GetAllocationTransferContextResponse.NotFound(_) => }
+    check("POST /registry/allocations/v1/{id}/choice-contexts/withdraw")(
+      allocationV1Client
+        .getAllocationWithdrawContext(
+          fakeCid,
+          allocation.v1.definitions.GetChoiceContextRequest(None),
+          headers,
+        )
+        .value
+        .futureValue
+    ) { case allocation.v1.GetAllocationWithdrawContextResponse.NotFound(_) => }
+    check("POST /registry/allocations/v1/{id}/choice-contexts/cancel")(
+      allocationV1Client
+        .getAllocationCancelContext(
+          fakeCid,
+          allocation.v1.definitions.GetChoiceContextRequest(None),
+          headers,
+        )
+        .value
+        .futureValue
+    ) { case allocation.v1.GetAllocationCancelContextResponse.NotFound(_) => }
+
+    // allocation v2
+    check("POST /registry/allocation/v2/settlement-factory")(
+      allocationV2Client
+        .getSettlementFactory(
+          allocation.v2.definitions.GetFactoryRequest(emptyChoiceArgs, None),
+          headers,
+        )
+        .value
+        .futureValue
+    ) { case allocation.v2.GetSettlementFactoryResponse.BadRequest(_) => }
+    check("POST /registry/allocations/v2/{id}/choice-contexts/withdraw")(
+      allocationV2Client
+        .getAllocationWithdrawContext(
+          fakeCid,
+          allocation.v2.definitions.GetChoiceContextRequest(None),
+          headers,
+        )
+        .value
+        .futureValue
+    ) { case allocation.v2.GetAllocationWithdrawContextResponse.NotFound(_) => }
+    check("POST /registry/allocations/v2/{id}/choice-contexts/cancel")(
+      allocationV2Client
+        .getAllocationCancelContext(
+          fakeCid,
+          allocation.v2.definitions.GetChoiceContextRequest(None),
+          headers,
+        )
+        .value
+        .futureValue
+    ) { case allocation.v2.GetAllocationCancelContextResponse.NotFound(_) => }
   }
 
 }
